重新命名了一些 标识符

+++ app/job/PushVideo.php
@@ -9,20 +9,18 @@ use FFMpeg\Format\Video\X264;
 use think\queue\Job;
 class PushVideo
 {
-    protected $ffmpeg;
-
     /**
      * @param Job $job
      * @param $data
      */
     public function fire(Job $job, $data) {
-        $isJobStillNeedToBeDone = $this->checkDatabaseToSeeIfJobNeedToBeDone($data);
-        $job->delete();
-        if(!$isJobStillNeedToBeDone){
+        $missionContinues = $this->checkMissionContinuesOrNot($data);
+        if(!$missionContinues){
+            $job->delete();
             return;
         }
-        $isJobDone = $this->liveStart($data);
-        if ($isJobDone) {
+        $jobDone = $this->pushStart($data);
+        if ($jobDone) {
             $job->delete();
         }else{
             if ($job->attempts() > 5) {
@@ -36,7 +34,7 @@ class PushVideo
      * @param array|mixed $data 将被推流的视频路径
      * @return boolean                 任务执行的结果
      */
-    private function checkDatabaseToSeeIfJobNeedToBeDone($data): bool
+    private function checkMissionContinuesOrNot($data): bool
     {
         return true;
     }
@@ -45,20 +43,20 @@ class PushVideo
      * @param string $videoUrl 将被推流的视频路径
      * @return bool
      */
-    private function liveStart(string $videoUrl): bool
+    private function pushStart(string $videoUrl): bool
     {
         echo "开始播放",PHP_EOL;
         $ffmpeg = FFMpeg::create([
             'ffmpeg.binaries'  => root_path() . "public/static/ffmpeg/ffmpeg.exe",
             'ffprobe.binaries' => root_path() . "public/static/ffmpeg/ffprobe.exe"
         ]);
-        $pushUrl = "rtmp://live-push.bilivideo.com/live-bvc/?streamname=live_188609215_9315200&key=ce264338a2392806e0634a40e63df74d&schedule=rtmp&pflag=1";
+        $pushPath = "rtmp://live-push.bilivideo.com/live-bvc/?streamname=live_188609215_9315200&key=ce264338a2392806e0634a40e63df74d&schedule=rtmp&pflag=1";
         $video = $ffmpeg->open($videoUrl);
         $format = new X264();
         $format->on('progress', function ($video, $format, $percentage) {
-            static $temp = 0;
-            if($temp != $percentage) {
-                $temp = $percentage;
+            static $RateProgress = 0;
+            if($RateProgress != $percentage) {
+                $RateProgress = $percentage;
                 echo "播放进度 $percentage %",PHP_EOL;
             }
         });
@@ -66,7 +64,7 @@ class PushVideo
             ->setInitialParameters(["-re","-i"])
             ->setAudioKiloBitrate(192)
             ->setAdditionalParameters(["-f","flv"]);
-        $video->save($format, $pushUrl);
+        $video->save($format, $pushPath);
         echo "结束播放",PHP_EOL;
         return false;
     }
